+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGPYTF7ENNQX3CJXCE96C0F2"
+        "01KGQ2CJCMF3MY1QMCNF02ESHK"
     )
     r"""The unique identifier of the data source"""