+++ packages/orq-rc/src/orq_ai_sdk/models/listidentitiesop.py
@@ -208,7 +208,7 @@ class ListIdentitiesData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-05T13:10:34.315Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-05T14:12:53.503Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")