+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGPYTF7HDR682CENM24GBFNP"
+        "01KGQ2CJCP8TVQ9EXYKN3HSK81"
     )
     r"""The unique identifier of the data source"""