+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -2193,7 +2193,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KGPYTEKYDKD1GDHNKT4FVMPS"
+    id: Optional[str] = "01KGQ2CHSW04CEK1A3QFVS2WFM"
 
     description: Optional[str] = None