+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kgpytepdfsh7x06cn8w44nte"
+        "conv_01kgq2chw12hka6bz5k26swh3p"
     )
     r"""Unique conversation identifier with `conv_` prefix."""