+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGPYTF7PGQ99STVPTED6S3ZB"
+        "01KGQ2CJCV27B4J4NPT24ZX98X"
     )
     r"""The unique identifier of the data source"""