+++ packages/orq-rc/RELEASES.md
@@ -5309,3 +5309,13 @@ Based on:
 - [python v4.3.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.29 - packages/orq-rc
+
+## 2026-02-05 14:15:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.705.0 (2.808.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.30] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.30 - packages/orq-rc